requirements.txt
Flask
tensorflow
numpy
Pillow
cassandra-driver
